+++ src/main.js
@@ -4,8 +4,9 @@ import {
   configForTsWithoutTypeChecking,
   configGlobalIgnore,
 } from './core/configs.js'
+import { createConfig } from './create-config.js'
 
-export const config = [configGlobalIgnore, configForJs, configForTsWithoutTypeChecking]
-export const configWithTypeChecking = [configGlobalIgnore, configForJs, configForTsWithTypeChecking]
-export { createConfig } from './create-config.js'
+export const config = createConfig()
+export const configWithTypeChecking = createConfig({ typeChecking: true })
+export { createConfig }
 export { configForJs, configForTsWithTypeChecking, configForTsWithoutTypeChecking }